requirements.txt
ament-clang-format==0.12.7
ament-copyright==0.12.7
ament-cppcheck==0.12.7
ament-cpplint==0.12.7
ament-flake8==0.12.7
ament-index-python==1.4.0
ament-lint==0.12.7
ament-lint-cmake==0.12.7
ament-package==0.14.0
ament-pep257==0.12.7
ament-uncrustify==0.12.7
ament-xmllint==0.12.7
argcomplete==3.1.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
cfgv==3.3.1
clang-format==14.0.6
comm==0.1.4
debugpy==1.6.7.post1
distlib==0.3.7
domain-coordinator==0.10.0
executing==1.2.0
filelock==3.12.2
gdown==4.7.1
identify==2.5.26
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
launch==1.0.4
launch-testing==1.0.4
launch-xml==1.0.4
launch-yaml==1.0.4
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
nest-asyncio==1.5.7
nodeenv==1.8.0
numpy @ file:///croot/numpy_and_numpy_base_1691091611330/work
osrf-pycommon==2.0.2
parso==0.8.3
pickleshare==0.7.5
pipx==1.2.0
platformdirs==3.9.1
pre-commit==3.3.3
prompt-toolkit==3.0.39
pure-eval==0.2.2
PySocks==1.7.1
python-dateutil==2.8.2
python-qt-binding==1.1.1
PyYAML==6.0.1
pyzmq==25.1.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
rcutils==5.1.3
resource-retriever==3.1.1
rosidl-cli==3.1.5
rpyutils==0.2.1
smclib==3.0.2
soupsieve==2.4.1
stack-data==0.6.2
tf-transformations==1.0.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
usd-core==23.8
userpath==1.9.0
virtualenv==20.24.2
wcwidth==0.2.6
xacro==2.0.8
